ghks027/spring_gangstargram#3 [조경환]  -gangstargram (게시글 작성) 게시글 작세 기능 - 본인 게시글만 삭제 가능 추가

+++ src/main/java/com/ganghwan/gangstargram/post/bo/PostBO.java
@@ -71,10 +71,16 @@ public class PostBO {
 	}
 	
 	// 게시글 삭제
-	public int deletePost(int postId) {
+	public int deletePost(int postId, int userId) {
+		
+		Post post = postDAO.selectPost(postId);
+		
+		// 본인 게시글만 삭제
+		if(post.getUserId() != userId) {
+			return 0;
+		}
 		
 		// 파일 삭제
-		Post post = postDAO.selectPost(postId);
 		FileManagerService.removeFile(post.getImage());
 		
 		// 댓글, 좋아요 삭제
